[TESY] Article 및 ArticleImage 테스트 코드 확장
src/test/java/com/teambind/articleserver/entity/ArticleImageTest.java: ArticleImagesId의 sequenceNum 타입 수정 (int → Long)
src/test/java/com/teambind/articleserver/entity/ArticleTest.java: 추가 및 삭제 관련 테스트 메서드 확장 (이미지/키워드 추가 및 삭제 테스트 구현)

참고: DB 동작 검증 위해 Rollback 설정 확인 필요

+++ src/test/java/com/teambind/articleserver/entity/ArticleImageTest.java
@@ -47,7 +47,7 @@ class ArticleImageTest {
 		
 		
 		List<ArticleImage> images = new ArrayList<>();
-		ArticleImagesId imagesId = new ArticleImagesId("article1", 0);
+		ArticleImagesId imagesId = new ArticleImagesId("article1", 0L);
 		ArticleImage articleImage = new ArticleImage();
 		articleImage.setId(imagesId);
 		articleImage.setImageUrl("test-url");
@@ -86,7 +86,7 @@ class ArticleImageTest {
 		
 		List<ArticleImage> images = new ArrayList<>();
 		for (int i = 0; i < 3; i++) {
-			ArticleImagesId imagesId = new ArticleImagesId("article1", i);
+			ArticleImagesId imagesId = new ArticleImagesId("article1", (long) i);
 			ArticleImage articleImage = new ArticleImage();
 			articleImage.setId(imagesId);
 			articleImage.setImageUrl("test-url-" + i);
